+++ app/src/main/java/com/jaimegc/covid19tracker/data/room/mapper/PojoMappers.kt
@@ -6,13 +6,6 @@ import com.jaimegc.covid19tracker.data.room.pojos.RegionAndOneStatsPojo
 import com.jaimegc.covid19tracker.data.room.pojos.RegionAndStatsPojo
 import com.jaimegc.covid19tracker.data.room.pojos.SubRegionAndOneStatsPojo
 import com.jaimegc.covid19tracker.data.room.pojos.SubRegionAndStatsPojo
-import com.jaimegc.covid19tracker.domain.model.ListRegionAndStats
-import com.jaimegc.covid19tracker.domain.model.ListSubRegionAndStats
-import com.jaimegc.covid19tracker.domain.model.RegionAndStats
-import com.jaimegc.covid19tracker.domain.model.SubRegionAndStats
-import com.jaimegc.covid19tracker.domain.model.toDomain
-import com.jaimegc.covid19tracker.domain.model.toRegionDomain
-import com.jaimegc.covid19tracker.domain.model.toSubRegionDomain
 
 fun List<CountryAndOneStatsPojo>.toPojoCountriesOrdered(): List<CountryAndStatsPojo> =
     this.groupBy { it.country }.let { mapCountries ->
@@ -55,23 +48,3 @@ fun List<SubRegionAndOneStatsPojo>.toPojoSubRegionsOrdered(): List<SubRegionAndS
         }
         listSubRegionAndStatsPojo
     }
-
-fun List<RegionAndStatsPojo>.toPojoRegionDomain(): ListRegionAndStats =
-    ListRegionAndStats(
-        map { entity ->
-            RegionAndStats(
-                region = entity.region!!.toDomain(),
-                stats = entity.stats.toRegionDomain()
-            )
-        }
-    )
-
-fun List<SubRegionAndStatsPojo>.toPojoSubRegionDomain(): ListSubRegionAndStats =
-    ListSubRegionAndStats(
-        map { entity ->
-            SubRegionAndStats(
-                subRegion = entity.subRegion!!.toDomain(),
-                stats = entity.stats.toSubRegionDomain()
-            )
-        }
-    )
